requirements.txt
openai
numpy==1.26.4
faiss-cpu
tqdm
pypdf
transformers==4.39.3
python-dotenv
streamlit